Create README.md
This program represents a kind of coffee vending machine. 
Different types of coffee can be selected via text input. 
The machine knows if there are still enough ingredients and can also be refilled. 
This project is part of the JetBrains Academy curriculum. 
The basic functions were created in the course. 
The structure of the program was rewritten to get a better structure and clarity.